Add CC BY SA to license options for site content
The MIT license is not a content and documentation license, per-se. It is true that the license /can/ cover documentation, but it is not an optimum choice. With the more widely accepted and used free culture Creative Commons licenses, the project signals that it is interesting in going away from being a content island and instead joining the mainland of content.

For example, putting this license forward AND posting it in the footer of the website (a future pull request) will signal to people that they are free to use/reuse/ modify blog content as they are with code.

However, I'll admit to not being sure the 4.0 Interrnational version of this license is the right choice. In fact, the Wikemedia Foundation legal team does not think the 4.0 is backwards compatible to the 3.0:

https://en.wikipedia.org/wiki/Wikipedia:FAQ/Copyright#Can_I_add_something_to_Wikipedia_that_I_got_from_somewhere_else.3F

As there is a large body of 3.0 content out there, I'm suggesting that one for now.

I think the BY SA options are philosophically aligned with the MIT, but am not a lawyer to say anything more.